Refactors PDF generation to use a shared utility
Extracts PDF generation logic into a reusable utility function
Adds loading overlay and error handling for consistent user experience
Updates multiple form classes to utilize the new utility

Improves code maintainability and reduces duplication
