+++ resources/js/staff/application-process-form-class/BaseApplicationForm.ts
@@ -10,6 +10,7 @@ import {
 import { processError } from '../../Utilities/error-handler-util';
 import '../../../css/pdf-loading.css';
 import { TableDataExtractor } from '../../Utilities/TableDataExtractor';
+import generatePDF from '../../Utilities/loading-overlay-pdf-generation';
 export type Action = 'edit' | 'view';
 
 /**
@@ -290,57 +291,7 @@ abstract class BaseApplicationForm {
                     this.generatePDFBtn?.attr('data-generated-url');
                 if (!generateUrl) throw new Error('Generate URL not found');
 
-                // Create loading overlay with classes instead of inline styles
-                const loadingOverlay = $('<div>', {
-                    class: 'pdf-loading-overlay',
-                });
-
-                const loadingSpinner = $('<div>', {
-                    class: 'spinner-border text-light',
-                    role: 'status',
-                });
-
-                const loadingText = $('<div>', {
-                    class: 'mt-3 text-light',
-                    text: 'Generating PDF, please wait...',
-                });
-
-                loadingOverlay.append(loadingSpinner, loadingText);
-                $('body').append(loadingOverlay);
-
-                try {
-                    // Open the PDF in a new window/tab
-
-                    // Fetch the PDF content
-                    const response = await fetch(generateUrl);
-
-                    if (!response.ok) {
-                        throw new Error(
-                            `Failed to generate PDF: ${response.statusText}`
-                        );
-                    }
-
-                    // Get the PDF blob
-                    const pdfBlob = await response.blob();
-
-                    // Create object URL
-                    const pdfUrl = URL.createObjectURL(pdfBlob);
-
-                    const newWindow = window.open('about:blank', '_blank');
-                    // Navigate the new window to the PDF
-                    if (newWindow) {
-                        newWindow.location.href = pdfUrl;
-                    }
-                } catch (error: any) {
-                    processError(
-                        `Error generating PDF for ${this.formName}:`,
-                        error,
-                        true
-                    );
-                } finally {
-                    // Remove loading overlay
-                    loadingOverlay.remove();
-                }
+                await generatePDF(generateUrl, this.formName);
             });
         } catch (error: any) {
             processError(
